+++ CHANGELOG.md
@@ -23,10 +23,15 @@ _Can be run using `npm run codetotal:beta`_
   - Add a report progress bar
   - Optimize new analysis dialog, drawer and linters list components' renders
   - Fix completed report receiving updates from ongoing analysis
+  
 - Back-End
   - Bug fix: SBOM packages not showing up in report page. Async parsing of packages information in SBOM module
   - Retry calls to pypi or npm in case first attempts are failing
   - Increase unit test coverage #88
+  - Fetch SBOM packages with concurrency of 10
+  - Fix issue when there is no SBOM package info
+  - SBOM results: Remove duplicate packages and sort them alphabetically (as a second criteria, after severity)
+  - Allow to override parallel number of SBOM packages fetches using env var CODETOTAL_SBOM_FETCH_PARALLEL_NB (default is now 30)
 
 ## [v0.5.0] - 2023-08-10